requirements.txt
PyQt5>=5.15.0
requests>=2.25.1
python-dotenv>=0.19.0
Flask>=2.0.1
Flask-CORS>=3.0.10
PyJWT>=2.1.0 